Send shaper to correct device

+++ mimikit/audios/features.py
@@ -243,7 +243,7 @@ class QuantizedSignal(Feature):
             ids = torch.from_numpy(shaper[1]).float().to(outputs)
             xvals = 2 * ids / ids[-1] - 1.0
             # unfortunately Interp1d does not let you select an axis - so process columns one by one
-            signal = torch.stack([Interp1d()(xvals, torch.from_numpy(shaper[0]).float().to(outputs), 2.0 * outputs[:,k] / (q_levels - 1) - 1.0) for k in range(outputs.shape[1])]).T
+            signal = torch.stack([Interp1d()(xvals, torch.from_numpy(shaper[0]).float().to(outputs), 2.0 * outputs[:,k].float() / (q_levels - 1) - 1.0) for k in range(outputs.shape[1])]).T
         elif sample_encoding == 'pcm':
             signal = 2.0 * outputs.float() / (q_levels - 1) - 1.0
         else:
